+++ store200/urls.py
@@ -2,4 +2,5 @@ from django.urls import path, include
 
 urlpatterns = [
         path('product', include('product.urls')),
+        path('order', include('order.urls'))
 ]
